Add Storage Drawers
Adds a 1x1 Storage Drawer GUI. I might go back to this to alter the arrow icons/the wood border, though as is this seems like a good template for the rest of them. I kept the wood the same color as it was, since it pops a lot more being a lighter tone (though if I redo the 'primitive' color schemes for GT Machine GUIs, I might take a tone from that later).