Add 2/4/Comp Drawers
Adds other variants of the Drawer GUI (really, this could be consolidated in the mod into one gui+some overlays. Not sure why that's not the case in the actual mod), and turned "drawers_1.psd" into "drawers_all.psd". When exporting, the slot layer names should be used for their respective layouts.