Fix Chisel, add Fish Trap
Fixed the incorrect locations for the last commit's chiselNEI GUI; I also added a retexture of the GT++ Fishtrap's interface. Since it's made of steel I used this as an excuse to use the gt-steel color palette (the black outline in the template is now very slightly tinted, which should avoid it being culled if rgb (0,0,0)->transparent.